Move utils to TileAndFuseUtils.[h|cpp]
