+++ compiler/src/iree/compiler/Codegen/Common/TileAndFuseUtils.cpp
@@ -5,10 +5,15 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "iree/compiler/Codegen/Common/TileAndFuseUtils.h"
+#include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenAttrs.h"
+#include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenDialect.h"
+#include "iree/compiler/Codegen/Interfaces/PartitionableLoopsInterface.h"
+#include "iree/compiler/Codegen/Utils/Utils.h"
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/Support/Casting.h"
 #include "llvm/Support/Debug.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
+#include "mlir/Dialect/SCF/Transforms/TileUsingInterface.h"
 #include "mlir/Dialect/Tensor/IR/Tensor.h"
 
 #include <cassert>
@@ -16,6 +21,8 @@
 #define DEBUG_TYPE "iree-codegen-common-tile-and-fuse-utils"
 
 namespace mlir::iree_compiler {
+
+#define CEILDIV(a, b) ((a + b - 1) / b)
 
 void fuseProducersOfSlices(RewriterBase &rewriter,
                            std::queue<Operation *> &worklist,
@@ -215,4 +222,213 @@ fuseConsumersIntoLoops(RewriterBase &rewriter, Operation *tiledOp,
   }
   return newFusionOpportunities;
 }
+
+FailureOr<TilingInfo> getTiledAndDistributionInfo(RewriterBase &rewriter,
+                                                  Operation *tilableOp) {
+  if (!tilableOp) {
+    // There is no lowering config. Return `null`.
+    return TilingInfo{nullptr, {}, {}};
+  }
+
+  IREE::Codegen::LoweringConfigAttrInterface tilableOpConfig =
+      getLoweringConfig(tilableOp);
+  if (!tilableOpConfig) {
+    return tilableOp->emitOpError("unable to find configuration of root op to "
+                                  "define workgroup count region");
+  }
+  auto tileSizes = llvm::map_to_vector(
+      tilableOpConfig.getWorkgroupTileSizes(),
+      [&](int64_t t) -> OpFoldResult { return rewriter.getIndexAttr(t); });
+  SmallVector<int64_t> interchange = tilableOpConfig.getWorkgroupInterchange();
+
+  // Avoid distributing unit-trip count loops.
+
+  // Set tile sizes for non-partitioned loops to zero.
+  if (auto partitionableLoopsInterface =
+          dyn_cast<PartitionableLoopsInterface>(tilableOp)) {
+    SmallVector<unsigned> partitionableLoops =
+        partitionableLoopsInterface.getPartitionableLoops(std::nullopt);
+    llvm::SmallDenseSet<unsigned> partitionableLoopsSet(
+        partitionableLoops.begin(), partitionableLoops.end());
+    OpFoldResult zero = rewriter.getIndexAttr(0);
+    for (auto loopId : llvm::seq<unsigned>(0, tileSizes.size())) {
+      if (partitionableLoopsSet.count(loopId)) {
+        continue;
+      }
+      tileSizes[loopId] = zero;
+    }
+  }
+
+  // Set tile sizes for full tiles to zero. This prevents single trip loops from
+  // being created, which can sometimes block certain cleanup patterns from
+  // applying during producer fusion.
+  if (auto tilingInterfaceOp = dyn_cast<TilingInterface>(tilableOp)) {
+    OpBuilder::InsertionGuard g(rewriter);
+    rewriter.setInsertionPoint(tilingInterfaceOp);
+    SmallVector<Range> bounds = tilingInterfaceOp.getIterationDomain(rewriter);
+    SmallVector<int64_t> staticLoopSizes;
+    SmallVector<Value> d;
+    for (Range bound : bounds) {
+      dispatchIndexOpFoldResult(bound.size, d, staticLoopSizes);
+    }
+    OpFoldResult zero = rewriter.getIndexAttr(0);
+    SmallVector<int64_t> tileSizesInt = tilableOpConfig.getWorkgroupTileSizes();
+    for (auto loopId : llvm::seq<unsigned>(0, tileSizesInt.size())) {
+      if (loopId < staticLoopSizes.size() &&
+          staticLoopSizes[loopId] == tileSizesInt[loopId]) {
+        tileSizes[loopId] = zero;
+      }
+    }
+  }
+
+  return TilingInfo{tilableOp, tileSizes, interchange};
+}
+
+SmallVector<Attribute>
+getDistributionMapping(MLIRContext *context, ArrayRef<OpFoldResult> tileSizes) {
+  SmallVector<Attribute> mapping;
+  mapping.reserve(tileSizes.size());
+  for (auto tileSize : llvm::reverse(tileSizes)) {
+    if (isZeroInteger(tileSize)) {
+      continue;
+    }
+    uint64_t currSize = mapping.size();
+    switch (currSize) {
+    case 0:
+    case 1:
+    case 2:
+      mapping.push_back(IREE::Codegen::WorkgroupMappingAttr::get(
+          context, IREE::Codegen::symbolizeWorkgroupId(currSize).value()));
+      break;
+    default:
+      mapping.push_back(IREE::Codegen::WorkgroupMappingAttr::get(
+          context, IREE::Codegen::WorkgroupId::IdZ, currSize - 2));
+    }
+  }
+  return llvm::to_vector(llvm::reverse(mapping));
+}
+
+bool areAllStaticLoopBounds(scf::ForallOp forallOp) {
+  for (auto [lb, ub, step] : llvm::zip_equal(forallOp.getMixedLowerBound(),
+                                             forallOp.getMixedUpperBound(),
+                                             forallOp.getMixedStep())) {
+    std::optional<int64_t> lbVal = getConstantIntValue(lb);
+    std::optional<int64_t> ubVal = getConstantIntValue(ub);
+    std::optional<int64_t> stepVal = getConstantIntValue(step);
+    if (!(lbVal && ubVal && stepVal)) {
+      return false;
+    }
+  }
+  return true;
+}
+
+static SmallVector<OpFoldResult>
+pruneDroppedLoops(ArrayRef<OpFoldResult> inputs,
+                  const llvm::SmallDenseSet<int> &droppedLoops) {
+  SmallVector<OpFoldResult> prunedInputs;
+  for (auto [index, input] : llvm::enumerate(inputs)) {
+    if (droppedLoops.contains(index)) {
+      continue;
+    }
+    prunedInputs.push_back(input);
+  }
+  return prunedInputs;
+}
+
+static SmallVector<Attribute>
+pruneDroppedLoops(ArrayRef<Attribute> inputs,
+                  const llvm::SmallDenseSet<int> &droppedLoops) {
+  SmallVector<IREE::Codegen::WorkgroupMappingAttr> droppedMappings;
+  SmallVector<Attribute> prunedAttrs;
+  for (auto [index, input] : llvm::enumerate(inputs)) {
+    if (droppedLoops.contains(index)) {
+      droppedMappings.push_back(
+          cast<IREE::Codegen::WorkgroupMappingAttr>(input));
+    } else {
+      prunedAttrs.push_back(input);
+    }
+  }
+  for (auto droppedMapping : droppedMappings) {
+    for (auto [index, prunedAttr] : llvm::enumerate(prunedAttrs)) {
+      auto prunedMappingAttr =
+          cast<IREE::Codegen::WorkgroupMappingAttr>(prunedAttr);
+      if (droppedMapping < prunedMappingAttr) {
+        prunedAttrs[index] =
+            IREE::Codegen::WorkgroupMappingAttr::getAttributeFromMappingId(
+                prunedAttr.getContext(), prunedMappingAttr.getMappingId() - 1);
+      }
+    }
+  }
+  return prunedAttrs;
+}
+
+FailureOr<scf::ForallOp> dropUnitDistributedDims(RewriterBase &rewriter,
+                                                 scf::ForallOp forallOp) {
+  SmallVector<OpFoldResult> mixedLbs = forallOp.getMixedLowerBound();
+  SmallVector<OpFoldResult> mixedUbs = forallOp.getMixedUpperBound();
+  SmallVector<OpFoldResult> mixedSteps = forallOp.getMixedStep();
+
+  // Find the index of loops to be dropped.
+  llvm::SmallDenseSet<int> droppedLoops;
+  for (auto [index, lb, ub, step] :
+       llvm::enumerate(mixedLbs, mixedUbs, mixedSteps)) {
+
+    std::optional<int64_t> lbVal = getConstantIntValue(lb);
+    std::optional<int64_t> ubVal = getConstantIntValue(ub);
+    std::optional<int64_t> stepVal = getConstantIntValue(step);
+
+    if (!(lbVal && ubVal && stepVal)) {
+      continue;
+    }
+
+    if (CEILDIV(ubVal.value() - lbVal.value(), stepVal.value()) == 1) {
+      droppedLoops.insert(index);
+    }
+  }
+  if (droppedLoops.empty()) {
+    return forallOp;
+  }
+
+  OpBuilder::InsertionGuard g(rewriter);
+  rewriter.setInsertionPoint(forallOp);
+  SmallVector<OpFoldResult> newLbs =
+      pruneDroppedLoops(ArrayRef<OpFoldResult>(mixedLbs), droppedLoops);
+  SmallVector<OpFoldResult> newUbs =
+      pruneDroppedLoops(ArrayRef<OpFoldResult>(mixedUbs), droppedLoops);
+  SmallVector<OpFoldResult> newSteps =
+      pruneDroppedLoops(ArrayRef<OpFoldResult>(mixedSteps), droppedLoops);
+  std::optional<ArrayAttr> newMapping;
+  if (auto currMapping = forallOp.getMapping()) {
+    SmallVector<Attribute> newMappingAttrs =
+        pruneDroppedLoops(currMapping.value().getValue(), droppedLoops);
+    newMapping = rewriter.getArrayAttr(newMappingAttrs);
+  }
+
+  Value zero = rewriter.create<arith::ConstantIndexOp>(forallOp.getLoc(), 0);
+  auto newForallOp = rewriter.create<scf::ForallOp>(
+      forallOp.getLoc(), newLbs, newUbs, newSteps, forallOp.getInits(),
+      newMapping, [](OpBuilder &, Location, ValueRange) {});
+
+  SmallVector<Value> argReplacements;
+  int newLoopBlockArgNum = 0;
+  auto newLoopBodyArgs = newForallOp.getInductionVars();
+  for (auto [index, oldBlockArg] :
+       llvm::enumerate(forallOp.getInductionVars())) {
+    if (droppedLoops.contains(index)) {
+      argReplacements.push_back(zero);
+    } else {
+      argReplacements.push_back(newLoopBodyArgs[newLoopBlockArgNum++]);
+    }
+  }
+  argReplacements.append(newForallOp.getRegionIterArgs().begin(),
+                         newForallOp.getRegionIterArgs().end());
+
+  Block *oldLoopBody = forallOp.getBody();
+  Block *newLoopBody = newForallOp.getBody();
+  rewriter.mergeBlocks(oldLoopBody, newLoopBody, argReplacements);
+
+  rewriter.replaceOp(forallOp, newForallOp.getResults());
+  return newForallOp;
+}
+
 } // namespace mlir::iree_compiler
